stage-3/detector/requirements.txt
pyyaml==6.0.1
flask==3.0.3
psutil==5.9.8
requests==2.31.0